added ffcv dataset,

+++ datasets/Aggregation_Dataset.py
@@ -39,7 +39,7 @@ class HistGen_Dataset(Dataset):
         tensor_file = torch.tensor(tensor_file["feats"][:]).to(torch.float32)
         label = torch.tensor(self.df.iloc[idx, 1]).type(torch.int64)
         censorship = torch.tensor(self.df.iloc[idx, 2]).type(torch.int64)
-        label_cont = torch.tensor(self.df.iloc[idx,3]).type(torch.int64)
+        label_cont = torch.tensor(self.df.iloc[idx,3]).type(torch.float32)
         return tensor_file, self.genomics_tensor[idx], censorship,  label,label_cont
 
         
@@ -72,7 +72,7 @@ class Gen_Dataset(Dataset):
         
         label = torch.tensor(self.df.iloc[idx, 1]).type(torch.int64)
         censorship = torch.tensor(self.df.iloc[idx, 2]).type(torch.int64)
-        label_cont = torch.tensor(self.df.iloc[idx,3]).type(torch.int64)
+        label_cont = torch.tensor(self.df.iloc[idx,3]).type(torch.float32)
         return self.genomics_tensor[idx], censorship,  label,label_cont
 
         
@@ -101,7 +101,7 @@ class Hist_Dataset(Dataset):
         
         label = torch.tensor(self.df.iloc[idx, 1]).type(torch.int64)
         censorship = torch.tensor(self.df.iloc[idx, 2]).type(torch.int64)
-        label_cont = torch.tensor(self.df.iloc[idx,3]).type(torch.int64)
+        label_cont = torch.tensor(self.df.iloc[idx,3]).type(torch.float32)
         return tensor_file, censorship,  label,label_cont
